src/requirements.txt
django
djangorestframework
psycopg2
pytz
pandas
